requirements.txt
numpy==2.3.1
opencv-python==4.11.0.86
pillow==11.2.1
